+++ sickrage/libs/adba/__init__.py
@@ -1,3 +1,4 @@
+#!/usr/bin/env python
 #
 # This file is part of aDBa.
 #
@@ -13,44 +14,30 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with aDBa.  If not, see <http://www.gnu.org/licenses/>.
-from __future__ import unicode_literals
-
 import threading
-import time
-from time import localtime, strftime
+from time import time, sleep, strftime, localtime
+from types import *
 
-from types import FunctionType, MethodType
-
-import sickrage
+from aniDBAbstracter import Anime, Episode
+from aniDBcommands import *
+from aniDBerrors import *
 from aniDBlink import AniDBLink
-from sickrage.indexers.adba.aniDBcommands import AnimeCommand, AuthCommand, \
-    BuddyAcceptCommand, BuddyAddCommand, BuddyDelCommand, BuddyDenyCommand, \
-    BuddyListCommand, BuddyStateCommand, EncryptCommand, EpisodeCommand, \
-    FileCommand, GroupCommand, GroupstatusCommand, LogoutCommand, \
-    MyListAddCommand, MyListCommand, MyListStatsCommand, NotifyAckCommand, \
-    NotifyAddCommand, NotifyCommand, NotifyDelCommand, NotifyGetCommand, \
-    NotifyListCommand, PingCommand, ProducerCommand, PushAckCommand, \
-    PushCommand, RandomAnimeCommand, SendMsgCommand, UptimeCommand, UserCommand, \
-    VersionCommand, VoteCommand, Notification
-from sickrage.indexers.adba.aniDBerrors import AniDBBannedError, \
-    AniDBCommandTimeoutError
 
 version = 100
 
-
 class Connection(threading.Thread):
-    def __init__(self, clientname='adba', server='api.anidb.info', port=9000, myport=9876, user=None, password=None,
-                 session=None, log=False, logPrivate=False, keepAlive=False):
-        super(Connection, self).__init__(name='AniDB')
+    def __init__(self, clientname='adba', server='api.anidb.info', port=9000, myport=9876, user=None, password=None, session=None, log=False, logPrivate=False, keepAlive=False):
+        super(Connection, self).__init__()
         # setting the log function
         self.logPrivate = logPrivate
-        if type(log) in (FunctionType, MethodType):  # if we get a function or a method use that.
+        if type(log) in (FunctionType, MethodType):# if we get a function or a method use that.
             self.log = log
-            self.logPrivate = True  # true means sensitive data will not be NOT be logged ... yeah i know oO
-        elif log:  # if it something else (like True) use the own print_log
+            self.logPrivate = True # true means sensitive data will not be NOT be logged ... yeah i know oO
+        elif log:# if it something else (like True) use the own print_log
             self.log = self.print_log
-        else:  # dont log at all
+        else:# dont log at all
             self.log = self.print_log_dummy
+
 
         self.link = AniDBLink(server, port, myport, self.log, logPrivate=self.logPrivate)
         self.link.session = session
@@ -59,7 +46,7 @@ class Connection(threading.Thread):
         self.clientver = version
 
         # from original lib 
-        self.mode = 1  # mode: 0=queue,1=unlock,2=callback
+        self.mode = 1    #mode: 0=queue,1=unlock,2=callback
 
         # to lock other threads out
         self.lock = threading.RLock()
@@ -78,13 +65,18 @@ class Connection(threading.Thread):
         self.counterAge = 0
 
     def print_log(self, data):
-        print(strftime("%Y-%m-%d %H:%M:%S", localtime(time.time())) + ": " + str(data))
+        print(strftime("%Y-%m-%d %H:%M:%S", localtime(time())) + ": " + str(data))
 
     def print_log_dummy(self, data):
         pass
 
     def stop(self):
         self.logout(cutConnection=True)
+        try:
+            self.join(1)
+        except:
+            pass
+
 
     def cut(self):
         self.link.stop()
@@ -96,20 +88,21 @@ class Connection(threading.Thread):
                 response.req.resp = None
                 response = self.handle(response.req, response.req.callback)
 
+
     def handle(self, command, callback):
 
         self.lock.acquire()
-        if self.counterAge < (time.time() - 120):  # the last request was older then 2 min reset delay and counter
+        if self.counterAge < (time() - 120): # the last request was older then 2 min reset delay and counter
             self.counter = 0
             self.link.delay = 2
-        else:  # something happend in the last 120 seconds
+        else: # something happend in the last 120 seconds
             if self.counter < 5:
-                self.link.delay = 2  # short term "A Client MUST NOT send more than 0.5 packets per second (that's one packet every two seconds, not two packets a second!)"
+                self.link.delay = 2 # short term "A Client MUST NOT send more than 0.5 packets per second (that's one packet every two seconds, not two packets a second!)"
             elif self.counter >= 5:
-                self.link.delay = 6  # long term "A Client MUST NOT send more than one packet every four seconds over an extended amount of time."
+                self.link.delay = 6 # long term "A Client MUST NOT send more than one packet every four seconds over an extended amount of time."
 
         if command.command not in ('AUTH', 'PING', 'ENCRYPT'):
-            self.counterAge = time.time()
+            self.counterAge = time()
             self.counter += 1
             if self.keepAlive:
                 self.authed()
@@ -121,11 +114,11 @@ class Connection(threading.Thread):
 
         self.log("handling(" + str(self.counter) + "-" + str(self.link.delay) + ") command " + str(command.command))
 
-        # make live request
+        #make live request
         command.authorize(self.mode, self.link.new_tag(), self.link.session, callback_wrapper)
         self.link.request(command)
 
-        # handle mode 1 (wait for response)
+        #handle mode 1 (wait for response)
         if self.mode == 1:
             command.wait_response()
             try:
@@ -145,7 +138,7 @@ class Connection(threading.Thread):
 
     def authed(self, reAuthenticate=False):
         self.lock.acquire()
-        authed = (self.link.session is not None)
+        authed = (self.link.session != None)
         if not authed and (reAuthenticate or self.keepAlive):
             self._reAuthenticate()
             authed = (self.link.session != None)
@@ -155,38 +148,38 @@ class Connection(threading.Thread):
     def _reAuthenticate(self):
         if self._username and self._password:
             self.log("auto re authenticating !")
-            try:
-                resp = self.auth(self._username, self._password)
-                if resp.rescode not in '500':
-                    return True
-            except Exception:
-                pass
-        return False
+            resp = self.auth(self._username, self._password)
+            if resp.rescode not in ('500'):
+                return True
+        else:
+            return False
 
     def _keep_alive(self):
-        self.lastKeepAliveCheck = time.time()
+        self.lastKeepAliveCheck = time()
         self.log("auto check !")
         # check every 30 minutes if the session is still valid
         # if not reauthenticate 
-        if self.lastAuth and time.time() - self.lastAuth > 1800:
+        if self.lastAuth and time() - self.lastAuth > 1800:
             self.log("auto uptime !")
-            self.uptime()  # this will update the self.link.session and will refresh the session if it is still alive
+            self.uptime() # this will update the self.link.session and will refresh the session if it is still alive
 
-            if self.authed():  # if we are authed we set the time
-                self.lastAuth = time.time()
-            else:  # if we aren't authed and we have the user and pw then reauthenticate
+            if self.authed(): # if we are authed we set the time
+                self.lastAuth = time()
+            else: # if we aren't authed and we have the user and pw then reauthenticate
                 self._reAuthenticate()
 
         # issue a ping every 20 minutes after the last package
         # this ensures the connection will be kept alive
-        if self.link.lastpacket and time.time() - self.link.lastpacket > 1200:
+        if self.link.lastpacket and time() - self.link.lastpacket > 1200:
             self.log("auto ping !")
             self.ping()
+
 
     def run(self):
         while self.keepAlive:
             self._keep_alive()
-            time.sleep(120)
+            sleep(120)
+
 
     def auth(self, username, password, nat=None, mtu=None, callback=None):
         """
@@ -199,32 +192,23 @@ class Connection(threading.Thread):
         mtu     - maximum transmission unit (max packet size) (default: 1400)
         
         """
+        self.log("ok1")
+        if self.keepAlive:
+            self.log("ok2")
+            self._username = username
+            self._password = password
+            if self.is_alive() == False:
+                self.log("You wanted to keep this thing alive!")
+                if self._iamALIVE == False:
+                    self.log("Starting thread now...")
+                    self.start()
+                    self._iamALIVE = True
+                else:
+                    self.log("not starting thread seams like it is already running. this must be a _reAuthenticate")
 
-        if not all([username, password]):
-            sickrage.srCore.srLogger.debug("AniDB username and/or password are not set. Aborting AniDB lookup.")
-            return False
 
-        authed = self.authed()
-        if not authed:
-            self.log("ok1")
-            if self.keepAlive:
-                self.log("ok2")
-                self._username = username
-                self._password = password
-                if not self.is_alive():
-                    self.log("You wanted to keep this thing alive!")
-                    if not self._iamALIVE:
-                        self.log("Starting thread now...")
-                        self.start()
-                        self._iamALIVE = True
-                    else:
-                        self.log("not starting thread seams like it is already running. this must be a _reAuthenticate")
-
-            self.lastAuth = time.time()
-            return self.handle(AuthCommand(username, password, 3, self.clientname, self.clientver, nat, 1, 'utf8', mtu),
-                               callback)
-
-        return authed
+        self.lastAuth = time()
+        return self.handle(AuthCommand(username, password, 3, self.clientname, self.clientver, nat, 1, 'utf8', mtu), callback)
 
     def logout(self, cutConnection=False, callback=None):
         """
@@ -232,7 +216,7 @@ class Connection(threading.Thread):
         
         """
         result = self.handle(LogoutCommand(), callback)
-        if (cutConnection):
+        if(cutConnection):
             self.cut()
         return result
 
@@ -442,7 +426,7 @@ class Connection(threading.Thread):
         """
         return self.handle(BuddyStateCommand(startat), callback)
 
-    def Anime(self, aid=None, aname=None, amask=-1, callback=None):
+    def anime(self, aid=None, aname=None, amask= -1, callback=None):
         """
         Get information about an anime
 
@@ -476,8 +460,7 @@ class Connection(threading.Thread):
         """
         return self.handle(EpisodeCommand(eid, aid, aname, epno), callback)
 
-    def file(self, fid=None, size=None, ed2k=None, aid=None, aname=None, gid=None, gname=None, epno=None, fmask=-1,
-             amask=0, callback=None):
+    def file(self, fid=None, size=None, ed2k=None, aid=None, aname=None, gid=None, gname=None, epno=None, fmask= -1, amask=0, callback=None):
         """
         Get information about a file
 
@@ -616,8 +599,7 @@ class Connection(threading.Thread):
 
         return self.handle(ProducerCommand(pid, pname), callback)
 
-    def mylist(self, lid=None, fid=None, size=None, ed2k=None, aid=None, aname=None, gid=None, gname=None, epno=None,
-               callback=None):
+    def mylist(self, lid=None, fid=None, size=None, ed2k=None, aid=None, aname=None, gid=None, gname=None, epno=None, callback=None):
         """
         Get information about your mylist
 
@@ -641,8 +623,7 @@ class Connection(threading.Thread):
         """
         return self.handle(MyListCommand(lid, fid, size, ed2k, aid, aname, gid, gname, epno), callback)
 
-    def mylistadd(self, lid=None, fid=None, size=None, ed2k=None, aid=None, aname=None, gid=None, gname=None, epno=None,
-                  edit=None, state=None, viewed=None, source=None, storage=None, other=None, callback=None):
+    def mylistadd(self, lid=None, fid=None, size=None, ed2k=None, aid=None, aname=None, gid=None, gname=None, epno=None, edit=None, state=None, viewed=None, source=None, storage=None, other=None, callback=None):
         """
         Add/Edit information to/in your mylist
 
@@ -684,9 +665,7 @@ class Connection(threading.Thread):
         -x    target all episodes upto x
         
         """
-        return self.handle(
-            MyListAddCommand(lid, fid, size, ed2k, aid, aname, gid, gname, epno, edit, state, viewed, source, storage,
-                             other), callback)
+        return self.handle(MyListAddCommand(lid, fid, size, ed2k, aid, aname, gid, gname, epno, edit, state, viewed, source, storage, other), callback)
 
     def mylistdel(self, lid=None, fid=None, aid=None, aname=None, gid=None, gname=None, epno=None, callback=None):
         """
@@ -788,6 +767,28 @@ class Connection(threading.Thread):
         
         """
         return self.handle(EncryptCommand(user, apipassword, type), callback)
+
+    def encoding(self, name, callback=None):
+        """
+        Change encoding used in messages
+
+        parameters:
+        name    - name of the encoding
+
+        structure of parameters:
+        name
+
+        comments:
+        DO NOT USE THIS!
+        utf8 is the only encoding which will support all the text in anidb responses
+        the responses have japanese, russian, french and probably other alphabets as well
+        even if you can't display utf-8 locally, don't change the server-client -connections encoding
+        rather, make python convert the encoding when you DISPLAY the text
+        it's better that way, let it go as utf8 to databases etc. because then you've the real data stored
+        
+        """
+        raise AniDBError, "pylibanidb sets the encoding to utf8 as default and it's stupid to use any other encoding. you WILL lose some data if you use other encodings, and now you've been warned. you will need to modify the code yourself if you want to do something as stupid as changing the encoding"
+        return self.handle(EncodingCommand(name), callback)
 
     def sendmsg(self, to, title, body, callback=None):
         """