+++ sickrage/libs/adba/aniDBerrors.py
@@ -1,3 +1,4 @@
+#!/usr/bin/env python
 #
 # This file is part of aDBa.
 #